+++ CSV/ZoneStarSizeII.csv
