+++ CSV/ZoneStarSizeVI.csv
