+++ Class/SystemGen.py
@@ -1,0 +1,36 @@
+"""
+Regroupe toutes les classes d'objets et les fonctions pour le fonctionnement de :System.py:
+Il y a différentes :Objet:
+    - :System: Systeme solaire
+    - :Star: Etoiles
+    - :Orbite: Orbite des Etoiles ne contenant pas les objets planètes mais certaines infos
+    - :Sattellite: Lunes et Astéroides
+    - :Planet: Toutes les planètes avec les détails
+"""
+from SystemFunc import *
+
+########################################################################################################################
+
+
+
+########################################################################################################################
+
+"""
+input("Test de la creation de systeme")
+again = True
+n = 0
+while n is not 10:
+    n += 1
+    S = System()
+    print(S)
+    S.clearorbit()
+
+
+input("Test suppression et creation orbit manuelle")
+for i in np.arange(10):
+    S = System(auto=False)
+    setattr(S,"Star1",Star())
+    s1 = S.Star1
+    s1.delOrbit("Orbit10")
+    s1.addOrbit(nom="Orbit1000")
+"""
